Provide a default thumbnail for all services
Previously we only set a thumbnail for services either with a category override or with an image specifically provided from ChurchSuite. This meant services with neither of these things had no thumbnail set, so YouTube would pick a frame for us.

We now explicitly set a thumbnail for all services.