Add privacy declarations back to LiveStream object
We previously removed these in favour of declaring them in the Video object. It turns out this breaks the update behaviour, as privacy is required in the status object for a livestream. Put them back.

As part of this, move the generation of the privacy string to the Service object, so we can properly isolate and test it.